PBL5-main/requirements.txt
flask==2.3.3
paho-mqtt==1.6.1
flask-cors==4.0.0
requests==2.31.0
